requirements.txt
python-telegram-bot==21.1.1
validators==0.28.1
requests
yt-dlp==2024.4.9 